refactor: Replace Handler with MouseUtils class
