+++ common/src/main/java/com/wynntils/core/components/Handlers.java
@@ -11,7 +11,6 @@ import com.wynntils.handlers.command.CommandHandler;
 import com.wynntils.handlers.container.ContainerQueryHandler;
 import com.wynntils.handlers.item.ItemHandler;
 import com.wynntils.handlers.labels.LabelHandler;
-import com.wynntils.handlers.mouseinput.MouseInputHandler;
 import com.wynntils.handlers.particle.ParticleHandler;
 import com.wynntils.handlers.scoreboard.ScoreboardHandler;
 import com.wynntils.handlers.tooltip.TooltipHandler;
@@ -24,7 +23,6 @@ public final class Handlers {
     public static final CommandHandler Command = new CommandHandler();
     public static final ContainerQueryHandler ContainerQuery = new ContainerQueryHandler();
     public static final ItemHandler Item = new ItemHandler();
-    public static final MouseInputHandler MouseInput = new MouseInputHandler();
     public static final LabelHandler Label = new LabelHandler();
     public static final ParticleHandler Particle = new ParticleHandler();
     public static final ScoreboardHandler Scoreboard = new ScoreboardHandler();
