+++ common/src/main/java/com/wynntils/utils/mc/MouseUtils.java
@@ -2,16 +2,14 @@
  * Copyright © Wynntils 2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.handlers.mouseinput;
+package com.wynntils.utils.mc;
 
-import com.wynntils.core.components.Handler;
-import com.wynntils.utils.mc.McUtils;
 import net.minecraft.network.protocol.game.ServerboundSwingPacket;
 import net.minecraft.network.protocol.game.ServerboundUseItemPacket;
 import net.minecraft.world.InteractionHand;
 
-public class MouseInputHandler extends Handler {
-    public void sendAttackInput(boolean reversed) {
+public final class MouseUtils {
+    public static void sendAttackInput(boolean reversed) {
         if (reversed) {
             McUtils.sendSequencedPacket(id -> new ServerboundUseItemPacket(
                     InteractionHand.MAIN_HAND,
@@ -24,11 +22,11 @@ public class MouseInputHandler extends Handler {
         }
     }
 
-    public void sendLeftClickInput() {
+    public static void sendLeftClickInput() {
         McUtils.sendPacket(new ServerboundSwingPacket(InteractionHand.MAIN_HAND));
     }
 
-    public void sendRightClickInput() {
+    public static void sendRightClickInput() {
         McUtils.sendSequencedPacket(id -> new ServerboundUseItemPacket(
                 InteractionHand.MAIN_HAND,
                 id,
